create a temp cryptography module with cryptography functions
