+++ crates/eip4844/src/trusted_setup.rs
@@ -1,7 +1,31 @@
 use bls12_381::{G1Point, G2Point};
 use serde::Deserialize;
 
+use crate::cryptography::{prover::CommitKey, verifier::VerificationKey};
+
 const TRUSTED_SETUP_JSON: &str = include_str!("../../eip7594/data/trusted_setup_4096.json");
+
+impl From<&TrustedSetup> for VerificationKey {
+    fn from(setup: &TrustedSetup) -> Self {
+        let g1_monomial = deserialize_g1_points(&setup.g1_monomial, SubgroupCheck::NoCheck);
+        let g2_monomial = deserialize_g2_points(&setup.g2_monomial, SubgroupCheck::NoCheck);
+        let gen_g1 = g1_monomial[0];
+        let gen_g2 = g2_monomial[0];
+        let tau_g2 = g2_monomial[1];
+        Self {
+            gen_g1,
+            gen_g2,
+            tau_g2,
+        }
+    }
+}
+
+impl From<&TrustedSetup> for CommitKey {
+    fn from(setup: &TrustedSetup) -> Self {
+        let g1_lagrange = deserialize_g1_points(&setup.g1_lagrange, SubgroupCheck::NoCheck);
+        Self { g1_lagrange }
+    }
+}
 
 #[derive(Deserialize, Debug, PartialEq, Eq)]
 pub struct TrustedSetup {
